store

+++ SpaccioDescans.Core/Application/Orders/Events/OrderCancelled.cs
@@ -3,7 +3,7 @@ using SpaccioDescans.Core.Domain.Stores;
 
 namespace SpaccioDescans.Core.Application.Orders.Events;
 
-public sealed record OrderCancelled(IEnumerable<long> ProductIds, int StoreId) : INotification;
+public sealed record OrderCancelled(IEnumerable<long> ProductIds, long StoreId) : INotification;
 
 public class OrderCancelledHandler : INotificationHandler<OrderCancelled>
 {
